+++ packages/ray-tracer-lib/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod camera;
-pub mod hitable;
 pub mod image;
 pub mod interval;
 pub mod materials;
