+++ packages/ray-tracer-lib/src/materials/material.rs
@@ -2,8 +2,8 @@ use glam::DVec3;
 
 use rand::rngs::ThreadRng;
 
+use crate::objects::HitRecord;
 use crate::ray::Ray;
-use crate::hitable::HitRecord;
 
 pub trait Material {
     fn scatter(
