+++ packages/ray-tracer-lib/src/objects/hitable.rs
