+++ packages/ray-tracer-lib/src/objects/mod.rs
@@ -1,3 +1,5 @@
 pub mod sphere;
+pub mod hitable;
 
+pub use hitable::*;
 pub use sphere::*;
