+++ packages/ray-tracer-lib/src/prelude.rs
@@ -1,7 +1,6 @@
 pub use crate::camera::*;
 pub use crate::ray::*;
 pub use crate::image::*;
-pub use crate::hitable::*;
 pub use crate::interval::*;
 pub use crate::materials::*;
 pub use crate::objects::*;
